chore: renamed examples
