--- basic/app.json
@@ -1,73 +1,0 @@
-{
-  "expo": {
-    "name": "basic",
-    "slug": "basic",
-    "version": "1.0.0",
-    "orientation": "portrait",
-    "icon": "./assets/icon.png",
-    "scheme": "basic",
-    "userInterfaceStyle": "light",
-    "newArchEnabled": true,
-    "splash": {
-      "image": "./assets/splash.png",
-      "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
-    },
-    "ios": {
-      "supportsTablet": true,
-      "bundleIdentifier": "expo.modules.shareextension.basic"
-    },
-     "android": {
-      "adaptiveIcon": {
-        "foregroundImage": "./assets/adaptive-icon.png",
-        "backgroundColor": "#ffffff"
-      },
-      "package": "expo.modules.shareextension.basic"
-    },
-    "web": {
-      "favicon": "./assets/favicon.png"
-    },
-    "plugins": [
-      [
-        "expo-font",
-        {
-          "fonts": ["./assets/fonts/Inter-Black.otf"]
-        }
-      ],
-      [
-        "../app.plugin.js",
-        {
-          "backgroundColor": {
-            "red": 255,
-            "green": 255,
-            "blue": 255,
-            "alpha": 0
-          },
-          "height": 500
-        }
-      ]
-    ],
-    "extra": {
-      "eas": {
-        "build": {
-          "experimental": {
-            "ios": {
-              "appExtensions": [
-                {
-                  "targetName": "basicShareExtension",
-                  "bundleIdentifier": "expo.modules.shareextension.basic.ShareExtension",
-                  "entitlements": {
-                    "com.apple.security.application-groups": [
-                      "group.expo.modules.shareextension.basic"
-                    ]
-                  }
-                }
-              ]
-            }
-          }
-        },
-        "projectId": "f729978a-6def-41c5-96b5-ae416ab2112f"
-      }
-    }
-  }
-}
