--- basic/index.share.js
@@ -1,5 +1,0 @@
-import { AppRegistry } from "react-native";
-
-import ShareExtension from "./src/ShareExtension";
-
-AppRegistry.registerComponent("shareExtension", () => ShareExtension);
