--- basic/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "basic",
-  "main": "index.js",
-  "version": "1.0.0",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo run:android",
-    "ios": "expo run:ios",
-    "prebuild": "expo prebuild -p ios --clean",
-    "lint": "expo lint"
-  },
-  "dependencies": {
-    "expo": "~54.0.13",
-    "expo-font": "~14.0.9",
-    "react": "19.1.0",
-    "react-native": "0.81.4",
-    "react-native-reanimated": "~4.1.1",
-    "react-native-worklets": "0.5.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.28.4",
-    "@babel/runtime": "^7.28.4",
-    "@types/react": "~19.1.0",
-    "eslint": "^9.25.0",
-    "eslint-config-expo": "~10.0.0",
-    "typescript": "~5.9.2"
-  },
-  "private": true,
-    "expo": {
-    "autolinking": {
-      "nativeModulesDir": ".."
-    }
-  }
-}
