+++ example/app.json
@@ -5,19 +5,19 @@
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
+    "scheme": "basic",
     "userInterfaceStyle": "light",
+    "newArchEnabled": true,
     "splash": {
       "image": "./assets/splash.png",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
-    "assetBundlePatterns": ["**/*"],
     "ios": {
       "supportsTablet": true,
       "bundleIdentifier": "expo.modules.shareextension.basic"
     },
-    "newArchEnabled": true,
-    "android": {
+     "android": {
       "adaptiveIcon": {
         "foregroundImage": "./assets/adaptive-icon.png",
         "backgroundColor": "#ffffff"
@@ -41,7 +41,7 @@
             "red": 255,
             "green": 255,
             "blue": 255,
-            "alpha": 0 // make the background transparent
+            "alpha": 0
           },
           "height": 500
         }
@@ -68,9 +68,6 @@
         },
         "projectId": "f729978a-6def-41c5-96b5-ae416ab2112f"
       }
-    },
-    "experiments": {
-      "reactCanary": true
     }
   }
 }
