+++ example/index.share.js
@@ -1,5 +1,5 @@
 import { AppRegistry } from "react-native";
 
-import ShareExtension from "./ShareExtension";
+import ShareExtension from "./src/ShareExtension";
 
 AppRegistry.registerComponent("shareExtension", () => ShareExtension);
