+++ example/package.json
@@ -1,29 +1,29 @@
 {
   "name": "basic",
-  "version": "1.0.0",
   "main": "index.js",
+  "version": "1.0.0",
   "scripts": {
     "start": "expo start",
     "android": "expo run:android",
     "ios": "expo run:ios",
-    "web": "expo start --web",
-    "eas-build-pre-install": "cd .. && npm install && npm run build && npm run build plugin",
-    "postinstall": "npx patch-package",
-    "prebuild": "expo prebuild -p ios --clean"
+    "prebuild": "expo prebuild -p ios --clean",
+    "lint": "expo lint"
   },
   "dependencies": {
-    "expo": "^53.0.9",
-    "expo-dev-client": "5.1.8",
-    "expo-splash-screen": "0.30.8",
-    "expo-status-bar": "2.2.3",
+    "expo": "~54.0.13",
+    "expo-font": "~14.0.9",
     "react": "19.1.0",
-    "react-native": "0.79.2",
-    "react-native-reanimated": "^3.17.5"
+    "react-native": "0.81.4",
+    "react-native-reanimated": "~4.1.1",
+    "react-native-worklets": "0.5.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.27.1",
-    "@types/react": "~19.1.4",
-    "typescript": "~5.8.3"
+    "@babel/core": "^7.28.4",
+    "@babel/runtime": "^7.28.4",
+    "@types/react": "~19.1.0",
+    "eslint": "^9.25.0",
+    "eslint-config-expo": "~10.0.0",
+    "typescript": "~5.9.2"
   },
   "private": true,
   "expo": {
