+++ example/tsconfig.json
@@ -3,8 +3,11 @@
   "compilerOptions": {
     "strict": true,
     "paths": {
-      "expo-share-extension": ["../src/index"],
-      "expo-share-extension/*": ["../src/*"]
+      "expo-share-extension": ["../src/index"],    
     }
-  }
+  },
+  "include": [
+    "**/*.ts",
+    "**/*.tsx"
+  ]
 }
